Refactor: Moved price fluctuation classes to strategy package
